+++ BE/src/main/java/codesquad/issuetracker/web/dto/GithubAccessToken.java
@@ -1,4 +1,4 @@
-package codesquad.issuetracker.oauth;
+package codesquad.issuetracker.web.dto;
 
 import com.fasterxml.jackson.databind.PropertyNamingStrategies;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
