+++ BE/src/main/java/codesquad/issuetracker/web/jwt/JwtFactory.java
@@ -1,4 +1,4 @@
-package codesquad.issuetracker.oauth;
+package codesquad.issuetracker.web.jwt;
 
 import com.auth0.jwt.algorithms.Algorithm;
 import org.springframework.beans.factory.annotation.Value;
